On line 68 in the CalendarUtil.java file, the calculation for determining the resulting year after adding a number of months to a date has an error.
Because the Gregorian calendar (used by the java.util.Date library)
represents the year 1900 as 0, a problem occurs when the resulting date
should be earlier than Jan 1900. The issue lies in integer division:

int resultYear = resultMonthCount / 12;

For example, if resultMonthCount is -1, then the resulting year will be
0 (1900) instead of -1 (1899) like it should be. I am submitting a patch
file which changes this line to the following:

int resultYear = (int) Math.floor(resultMonthCount / 12.0);

This will ensure that the number will result in the correct quotient for
both positive, zero and negative values of resultMonthCount.

Review-Link: http://gwt-code-reviews.appspot.com/1898803/

Review by: skybrian@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@11564 8db76d5a-ed1c-0410-87a9-c151d255dfc7