Added gwt.codeserver.port system property to RemoteServiceServlet
This property is the port of a Super Dev Mode code server running
on localhost. When a serialization policy isn't found locally, the
servlet attempts to download it from the code server.

API change: added two protected methods that can be used to customize
the behavior of this feature:

getCodeServerPolicyUrl() controls whether it's enabled and the URL
loadPolicyFromCodeServer() can be overridden to change the implementation

Change-Id: Iaf2366cb94297e738cf50bb9d8038fef66b89722
Review-Link: https://gwt-review.googlesource.com/#/c/2341/

Review by: mdempsky@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@11571 8db76d5a-ed1c-0410-87a9-c151d255dfc7